+++ db/migrate/20120106015347_add_extracted_file_target.rb
@@ -1,9 +1,9 @@
 class AddExtractedFileTarget < ActiveRecord::Migration  
   def self.up
-    add_column :data_source_downloads, :file_to_extract, :string
+    add_column :data_source_downloads, :file_to_extract, :string unless column_exists? :data_source_downloads, :file_to_extract
   end
 
   def self.down
-    remove_column :data_source_downloads, :file_to_extract
+    remove_column :data_source_downloads, :file_to_extract if column_exists? :data_source_downloads, :file_to_extract
   end
 end
