+++ db/migrate/20120823023412_remove_log_field.rb
@@ -1,9 +1,9 @@
 class RemoveLogField < ActiveRecord::Migration
   def self.up
-    remove_column :data_source_downloads, :download_log
+    remove_column :data_source_downloads, :download_log if column_exists? :data_source_downloads, :download_log
   end
 
   def self.down
-    add_column :data_source_downloads, :download_log, :text, :limit => 4294967295
+    add_column :data_source_downloads, :download_log, :text, :limit => 4294967295 unless column_exists? :data_source_downloads, :download_log
   end
 end
