+++ db/migrate/20130301003211_drop_original_last_run.rb
@@ -1,9 +1,9 @@
 class DropOriginalLastRun < ActiveRecord::Migration
   def self.up
-    remove_column :data_sources, :last_run
+    remove_column :data_sources, :last_run if column_exists? :data_sources, :last_run
   end
 
   def self.down
-    add_column :data_sources, :last_run, :datetime
+    add_column :data_sources, :last_run, :datetime unless column_exists? :data_sources, :last_run
   end
 end
