+++ db/migrate/20140301004236_add_data_source_priority.rb
@@ -1,9 +1,9 @@
 class AddDataSourcePriority < ActiveRecord::Migration
   def self.up
-    add_column :data_sources, :priority, :integer, :default => 100    
+    add_column :data_sources, :priority, :integer, :default => 100 unless column_exists? :data_sources, :priority
   end
 
   def self.down
-    remove_column :data_sources, :priority
+    remove_column :data_sources, :priority if column_exists? :data_sources, :priority
   end
 end
